+++ sdk/attestation/azure-security-attestation/samples/policy/reset_sealed_policy.cpp
@@ -60,8 +60,8 @@ int main()
         GetEnvHelper::GetEnv("AZURE_TENANT_ID"),
         GetEnvHelper::GetEnv("AZURE_CLIENT_ID"),
         GetEnvHelper::GetEnv("AZURE_CLIENT_SECRET"));
-    AttestationAdministrationClient const adminClient(
-        AttestationAdministrationClient::Create(endpoint, credential, clientOptions));
+    std::shared_ptr<AttestationAdministrationClient const> adminClient(
+        AttestationAdministrationClientFactory::Create(endpoint, credential, clientOptions));
 
     std::string const signingKey(GetEnvHelper::GetEnv("ISOLATED_SIGNING_KEY"));
     std::string const signingCert(GetEnvHelper::GetEnv("ISOLATED_SIGNING_CERTIFICATE"));
@@ -76,7 +76,7 @@ int main()
     resetOptions.SigningKey = AttestationSigningKey{pemSigningKey, pemSigningCert};
 
     Azure::Response<AttestationToken<PolicyResult>> const resetResult
-        = adminClient.ResetAttestationPolicy(AttestationType::SgxEnclave, resetOptions);
+        = adminClient->ResetAttestationPolicy(AttestationType::SgxEnclave, resetOptions);
 
     if (resetResult.Value.Body.PolicyResolution == PolicyModification::Updated)
     {