+++ sdk/attestation/azure-security-attestation/test/ut/tpmattestation_test.cpp
@@ -98,7 +98,7 @@ namespace Azure { namespace Security { namespace Attestation { namespace Test {
       std::shared_ptr<Azure::Core::Credentials::TokenCredential> credential
           = CreateClientSecretCredential(
               GetEnv("AZURE_TENANT_ID"), GetEnv("AZURE_CLIENT_ID"), GetEnv("AZURE_CLIENT_SECRET"));
-      return AttestationClient::CreatePointer(GetInstanceUri(instanceType), credential, options);
+      return AttestationClientFactory::Create(GetInstanceUri(instanceType), credential, options);
     }
 
     std::unique_ptr<AttestationAdministrationClient> CreateAdminClient(InstanceType instanceType)
@@ -110,7 +110,7 @@ namespace Azure { namespace Security { namespace Attestation { namespace Test {
       std::shared_ptr<Azure::Core::Credentials::TokenCredential> credential
           = CreateClientSecretCredential(
               GetEnv("AZURE_TENANT_ID"), GetEnv("AZURE_CLIENT_ID"), GetEnv("AZURE_CLIENT_SECRET"));
-      return AttestationAdministrationClient::CreatePointer(
+      return AttestationAdministrationClientFactory::Create(
           GetInstanceUri(instanceType), credential, options);
     }
   };